refactor deterministic sim
